Add Delete Repository function, rearrange project toolbar
Deleting the repository will remove it from the project, delete local files, but NOT delete the remote branch. This mirrors the way other tools (like GoLand/IntelliJ) usually work.

The side effect of this is that if you delete a repo from a project and then re-add it, you may have to manually do a force push or pull to get things synced again.